refactor: created data middleware
